backend/modules/payroll/tests/requirements-test.txt
# backend/modules/payroll/tests/requirements-test.txt

# Testing frameworks
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # Parallel test execution

# Test utilities
factory-boy==3.3.0  # Test data factories
faker==20.1.0  # Fake data generation
freezegun==1.2.2  # Time mocking
responses==0.24.1  # HTTP response mocking

# Code quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Database testing
pytest-postgresql==5.0.0
sqlalchemy-utils==0.41.1

# Performance testing
pytest-benchmark==4.0.0
memory-profiler==0.61.0
psutil==5.9.6

# Reporting
pytest-html==4.1.1
pytest-json-report==1.5.0
allure-pytest==2.13.2